ci hecho y corrección de pruebas
