+++ .github/workflows/ci.yml
@@ -1,0 +1,32 @@
+name: Pruebas de Integración
+
+on:
+  push:
+    branches: [ main ]
+  pull_request:
+    branches: [ main ]
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: 1. Clonar el repositorio
+        uses: actions/checkout@v4
+
+      - name: 2. Configurar pnpm
+        uses: pnpm/action-setup@v4
+        with:
+          version: 9
+
+      - name: 3. Configurar Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version: '20'
+          cache: 'pnpm'
+
+      - name: 4. Instalar dependencias
+        run: pnpm install
+
+      - name: 5. Ejecutar pruebas de integración
+        run: pnpm test:ci
