--- .github/workflows/deploy.yml
@@ -1,2 +1,0 @@
-name: ci-deploy
-on: { push: { branches: [ main ] } }
