+++ .wrangler/tmp/bundle-GLBmnS/middleware-insertion-facade.js
@@ -1,8 +1,8 @@
-				import worker, * as OTHER_EXPORTS from "C:\\DevOps\\Desarrollos\\ecv-shortener\\.wrangler\\tmp\\pages-8ePkwp\\functionsWorker-0.16915985605163852.mjs";
+				import worker, * as OTHER_EXPORTS from "C:\\DevOps\\Desarrollos\\ecv-shortener\\.wrangler\\tmp\\pages-t2RGbj\\functionsWorker-0.5146835368302396.mjs";
 				import * as __MIDDLEWARE_0__ from "C:\\DevOps\\Desarrollos\\ecv-shortener\\node_modules\\.pnpm\\wrangler@4.33.1_@cloudflare+workers-types@4.20250828.0\\node_modules\\wrangler\\templates\\middleware\\middleware-ensure-req-body-drained.ts";
 import * as __MIDDLEWARE_1__ from "C:\\DevOps\\Desarrollos\\ecv-shortener\\node_modules\\.pnpm\\wrangler@4.33.1_@cloudflare+workers-types@4.20250828.0\\node_modules\\wrangler\\templates\\middleware\\middleware-miniflare3-json-error.ts";
 
-				export * from "C:\\DevOps\\Desarrollos\\ecv-shortener\\.wrangler\\tmp\\pages-8ePkwp\\functionsWorker-0.16915985605163852.mjs";
+				export * from "C:\\DevOps\\Desarrollos\\ecv-shortener\\.wrangler\\tmp\\pages-t2RGbj\\functionsWorker-0.5146835368302396.mjs";
 				const MIDDLEWARE_TEST_INJECT = "__INJECT_FOR_TESTING_WRANGLER_MIDDLEWARE__";
 				export const __INTERNAL_WRANGLER_MIDDLEWARE__ = [
