+++ .wrangler/tmp/bundle-GLBmnS/middleware-loader.entry.ts
@@ -3,12 +3,12 @@
 // export dynamically through wrangler, or we can potentially let users directly
 // add them as a sort of "plugin" system.
 
-import ENTRY, { __INTERNAL_WRANGLER_MIDDLEWARE__ } from "C:\\DevOps\\Desarrollos\\ecv-shortener\\.wrangler\\tmp\\bundle-WG1XcG\\middleware-insertion-facade.js";
+import ENTRY, { __INTERNAL_WRANGLER_MIDDLEWARE__ } from "C:\\DevOps\\Desarrollos\\ecv-shortener\\.wrangler\\tmp\\bundle-GLBmnS\\middleware-insertion-facade.js";
 import { __facade_invoke__, __facade_register__, Dispatcher } from "C:\\DevOps\\Desarrollos\\ecv-shortener\\node_modules\\.pnpm\\wrangler@4.33.1_@cloudflare+workers-types@4.20250828.0\\node_modules\\wrangler\\templates\\middleware\\common.ts";
-import type { WorkerEntrypointConstructor } from "C:\\DevOps\\Desarrollos\\ecv-shortener\\.wrangler\\tmp\\bundle-WG1XcG\\middleware-insertion-facade.js";
+import type { WorkerEntrypointConstructor } from "C:\\DevOps\\Desarrollos\\ecv-shortener\\.wrangler\\tmp\\bundle-GLBmnS\\middleware-insertion-facade.js";
 
 // Preserve all the exports from the worker
-export * from "C:\\DevOps\\Desarrollos\\ecv-shortener\\.wrangler\\tmp\\bundle-WG1XcG\\middleware-insertion-facade.js";
+export * from "C:\\DevOps\\Desarrollos\\ecv-shortener\\.wrangler\\tmp\\bundle-GLBmnS\\middleware-insertion-facade.js";
 
 class __Facade_ScheduledController__ implements ScheduledController {
 	readonly #noRetry: ScheduledController["noRetry"];
