+++ package.json
@@ -3,14 +3,15 @@
   "private": true,
   "packageManager": "pnpm@9.0.0",
   "scripts": {
-    "dev:pages": "wrangler pages dev apps/web/dist --compatibility-date=2024-11-01",
+    "dev:pages": "wrangler pages dev apps/web/dist --kv=LINKS --persist-to=.wrangler/state --compatibility-date=2024-11-01",
     "dev": "pnpm --filter @ecv/web dev",
     "build": "pnpm -r run build",
     "typecheck": "pnpm -r run typecheck",
     "lint": "eslint . --ext .ts,.tsx",
     "test": "pnpm --filter @ecv/domain test",
     "preview": "pnpm --filter @ecv/web preview",
-    "test:integration": "vitest run --config vitest.config.integration.ts"
+    "test:integration": "vitest run --config vitest.config.integration.ts",
+    "test:ci": "start-server-and-test dev:pages http://localhost:8788 test:integration"
   },
   "devDependencies": {
     "@cloudflare/workers-types": "^4.20240821.0",
@@ -21,6 +22,7 @@
     "@typescript-eslint/parser": "^8.0.1",
     "eslint": "^9.9.0",
     "prettier": "^3.3.3",
+    "start-server-and-test": "^2.0.13",
     "typescript": "^5.5.4",
     "vitest": "^3.2.4",
     "wrangler": "^4.32.0",
